chore: extract `findEsbuildConfig` into own module
for test orchestration
